Remove LegacyDifficultyCalculator
